product-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
